Relocate BooleanWidget to Widgets directory
