+++ packages/airview-cms/src/features/cms/widgets/index.js
@@ -1,2 +1,3 @@
 export { DynamicWidget } from "./dynamic-widget";
 export { TextFieldWidget } from "./text-field-widget";
+export { BooleanWidget } from "./boolean-widget";
